+++ ModelForge/routers/finetuning_router.py
@@ -9,10 +9,10 @@ from fastapi import Request
 from starlette.responses import JSONResponse
 from pydantic import BaseModel, field_validator
 
-from ..utilities.CausalLLMTuner import CausalLLMFinetuner
-from ..utilities.QuestionAnsweringTuner import QuestionAnsweringTuner
-from ..utilities.Seq2SeqLMTuner import Seq2SeqFinetuner
-from ..utilities.model_validator import ModelValidator
+from utilities.settings_managers.CausalLLMTuner import CausalLLMFinetuner
+from utilities.finetuning.QuestionAnsweringTuner import QuestionAnsweringTuner
+from utilities.finetuning.Seq2SeqLMTuner import Seq2SeqFinetuner
+from utilities.hardware_detection.model_validator import ModelValidator
 
 from ..globals.globals_instance import global_manager
