+++ ModelForge/utilities/settings_managers/CausalLLMTuner.py
@@ -1,10 +1,10 @@
 import torch
 from datasets import load_dataset
 from trl import SFTTrainer, SFTConfig
-from peft import LoraConfig, get_peft_model, TaskType, PeftModel
+from peft import LoraConfig, get_peft_model, TaskType
 from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig
 from typing import Dict, Optional
-from .Finetuner import Finetuner
+from utilities.finetuning.Finetuner import Finetuner
 import os
 from huggingface_hub import errors as hf_errors
 import traceback
